Fixed issue with the slow implementation warning
Currently there is a bug if we go into the last else:
```
  File "/storage/ekrivov/AutoAWQ/awq/modules/linear/gemm.py", line 270, in forward
    out = WQLinearMMFunction.apply(
          ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/runner/miniforge3/envs/triton/lib/python3.12/site-packages/torch/autograd/function.py", line 578, in apply
    return super().apply(*args, **kwargs)  # type: ignore[misc]
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/storage/ekrivov/AutoAWQ/awq/modules/linear/gemm.py", line 70, in forward
    if not user_has_been_warned:
           ^^^^^^^^^^^^^^^^^^^^
UnboundLocalError: cannot access local variable 'user_has_been_warned' where it is not associated with a value
```

We need to make it global before we can modify it. And we have to make it global before the first read access.